Update Functiondoestimation.java
This is an attempt to suppress logging using the Java Logger class. In many cases, TensorFlow is not using Java's built-in logging system, so attempts to suppress logs in this way might not be effective. The TensorFlow library has its own internal logging system and reads the TF_CPP_MIN_LOG_LEVEL environment variable directly from the system when the library is loaded.