add optimization to dss catalog read to not pull ts extents
in heclib, the times were only queried on small DSS6 files. For heclib7 and DSS7 the extents are retrieved along with the catalog, making the operation fast. For heclib7 and DSS6, the library was changed to always query the ts extents regardless of size, which is very slow for very large DSS files.

This changes makes the same calls that were done with heclib6 by only conditionally querrying the times.